backend-java/ai-service/requirements.txt
# AI Service Dependencies
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
jinja2==3.1.3
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.8
python-multipart==0.0.6
aiofiles==23.2.1

